requirements.txt
streamlit
opencv-python
numpy
Pillow
import cv2